chore: remove legacy code
