+++ mypy.ini
@@ -42,7 +42,7 @@ ignore_missing_imports      = True
 [mypy-momento.cache_operation_responses]
 disallow_any_expr           = False
 
-[mypy-momento.errors.cache_service_errors_converter]
+[mypy-momento.errors.error_converter]
 disallow_any_expr           = False
 
 [mypy-momento.errors.error_details]
