+++ src/momento/errors/__init__.py
@@ -1,18 +1,3 @@
-from .all_errors import (
-    AlreadyExistsError,
-    AuthenticationError,
-    BadRequestError,
-    CancelledError,
-    ClientSdkError,
-    InternalServerError,
-    InvalidArgumentError,
-    LimitExceededError,
-    MomentoServiceError,
-    NotFoundError,
-    PermissionError,
-    SdkError,
-    TimeoutError,
-)
 from .error_details import (
     MomentoErrorCode,
     MomentoErrorTransportDetails,
@@ -38,4 +23,4 @@ from .exceptions import (
 
 # NB: since this module imports from sibling modules, it must be at the bottom
 # to avoid circular imports
-from .cache_service_errors_converter import new_convert
+from .error_converter import convert_error
