+++ src/momento/responses/__init__.py
@@ -1,10 +1,14 @@
 from .control import (
     CreateCacheResponse,
     CreateCacheResponseBase,
+    CreateSigningKeyResponse,
     DeleteCacheResponse,
     DeleteCacheResponseBase,
     ListCachesResponse,
     ListCachesResponseBase,
+    ListSigningKeysResponse,
+    RevokeSigningKeyResponse,
+    SigningKey,
 )
 from .scalar_data import (
     CacheDeleteResponse,
