+++ src/momento/responses/control/__init__.py
@@ -1,3 +1,9 @@
 from .create_cache import CreateCacheResponse, CreateCacheResponseBase
 from .delete_cache import DeleteCacheResponse, DeleteCacheResponseBase
 from .list_caches import ListCachesResponse, ListCachesResponseBase
+from .signing_keys import (
+    CreateSigningKeyResponse,
+    ListSigningKeysResponse,
+    RevokeSigningKeyResponse,
+    SigningKey,
+)
