+++ src/momento/simple_cache_client.py
@@ -29,23 +29,33 @@ except ImportError as e:
     raise e
 
 import momento._momento_endpoint_resolver as endpoint_resolver
-from momento.cache_operation_types import (
-    CreateSigningKeyResponse,
-    ListSigningKeysResponse,
-    RevokeSigningKeyResponse,
-)
 from momento.responses import (
     CacheDeleteResponseBase,
     CacheGetResponseBase,
     CacheSetResponseBase,
     CreateCacheResponseBase,
+    CreateSigningKeyResponse,
     DeleteCacheResponseBase,
     ListCachesResponseBase,
+    ListSigningKeysResponse,
+    RevokeSigningKeyResponse,
 )
 
 
 class SimpleCacheClient:
-    """Synchronous Simple Cache Client"""
+    """Async Simple Cache Client"""
+
+    # For high load, we might get better performance with multiple clients, because the server is
+    # configured to allow a max of 100 streams per connection.  In the javascript SDK, multiple
+    # clients resulted in an obvious performance improvement.  However, in the python SDK I have
+    # not yet been able to observe a clear benefit.  So for now, we are putting the plumbing in
+    # place so that we can more easily test performance with multiple connections in the future,
+    # but we are leaving the default value set to 1.
+    #
+    # We are hard-coding the value for now, because we haven't yet designed the API for
+    # users to use to configure tunables:
+    # https://github.com/momentohq/dev-eco-issue-tracker/issues/85
+    _NUM_CLIENTS = 1
 
     def __init__(
         self,
