+++ tests/test_momento_signer.py
@@ -2,7 +2,7 @@ import jwt
 import pytest
 from jwt.api_jwk import PyJWK
 
-from momento.errors import InvalidArgumentError
+from momento.errors import InvalidArgumentException
 from momento.momento_signer import CacheOperation, MomentoSigner, SigningRequest
 from tests.utils import uuid_str
 
@@ -129,17 +129,17 @@ def test_es_no_alg():
 
 
 def test_empty_jwk_json_string():
-    with pytest.raises(InvalidArgumentError):
+    with pytest.raises(InvalidArgumentException):
         MomentoSigner("")
 
 
 def test_nothing_jwk_json_string():
-    with pytest.raises(InvalidArgumentError):
+    with pytest.raises(InvalidArgumentException):
         MomentoSigner("{}")
 
 
 def test_incomplete_jwk_json_string():
-    with pytest.raises(InvalidArgumentError):
+    with pytest.raises(InvalidArgumentException):
         MomentoSigner('{"alg":"foo","kid":"bar","kty":"RSA"}')
 
 
@@ -179,7 +179,7 @@ def test_create_presigned_url_for_set():
 
 
 def test_create_presigned_url_for_set_missing_ttl():
-    with pytest.raises(InvalidArgumentError):
+    with pytest.raises(InvalidArgumentException):
         MomentoSigner(_RSA_256_JWK).create_presigned_url(
             "example.com",
             SigningRequest(
